proper exception type
